break up bmpshower into image, surface and volume viewer widget files (renamed classes too)
