+++ TestRunner/CMakeLists.txt
@@ -10,8 +10,6 @@
 IF(ISEG_BUILD_TESTING)
 	USE_BOOST()
 
-	MESSAGE("MY_EXTERNAL_LINK_LIBRARIES ${MY_EXTERNAL_LINK_LIBRARIES}")
-
 	ADD_EXECUTABLE(TestRunner TestRunner.cpp)
 	TARGET_LINK_LIBRARIES(TestRunner ${MY_EXTERNAL_LINK_LIBRARIES})
 	VS_SET_PROPERTY(TestRunner "TestSuite")
